# packages/api-client/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Shelter.Host | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/platform/session/login": {
      "post": {
        "tags": [
          "platform"
        ],
        "operationId": "LoginPlatformSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/platform/session/organization/{organizationId}": {
      "post": {
        "tags": [
          "platform"
        ],
        "operationId": "SelectPlatformSessionOrganization",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}